requirements.txt
# ============================================================
# CS饰品指数分析工具 - Python依赖
# 安装: pip install -r requirements.txt
# Python版本要求: >= 3.9
# ============================================================

# ---- 核心依赖 ----
requests>=2.31.0           # HTTP请求（调用CSQAQ API / 企业微信Webhook）
PyYAML>=6.0.1              # YAML配置文件解析

# ---- UI 框架 ----
PyQt5>=5.15.9              # 主UI框架（跨平台GUI）

# ---- 数据处理与可视化 ----
numpy>=1.24.0              # 数值计算（MA均线等指标）
pandas>=2.0.0              # 数据处理与时序分析
matplotlib>=3.7.0          # K线图绑制（含 candlestick_ohlc / mplfinance）
mplfinance>=0.12.10b0      # 专业的金融K线图绘制库
Pillow>=10.0.0             # 图像处理（K线图导出/压缩）

# ---- 日志 ----
# (logging为Python内置模块，无需额外安装)

# ---- 开发工具（可选）----
# pytest>=7.4.0            # 单元测试
# black>=23.0.0            # 代码格式化
# flake8>=6.0.0            # 代码风格检查
